% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.s.miss.R
\name{R.s.miss}
\alias{R.s.miss}
\title{Surrogate marker validation with missing values
This function calculates the proportion of treatment effect on the primary outcome explained by the treatment effect on a surrogate marker, correcting for missing values in the surrogate marker. This function is intended to be used for a fully observed continuous outcome. The user must specify what type of estimation they would like (parametric or nonparametric estimation of the proportion explained, denoted by R) and what estimator they would like (see below for details). Missing values can be handled through inverse probability weighting (IPW) or semiparametric sieve maximum likelihood estimation (SMLE).}
\usage{
R.s.miss(
  sone,
  szero,
  yone,
  yzero,
  wone = NULL,
  wzero = NULL,
  type = "robust",
  max.it = 10000,
  tol = 0.001,
  conf.int = FALSE,
  ipw.formula = NULL,
  orig.smle = TRUE
)
}
\arguments{
\item{sone}{numeric vector; surrogate marker for treated observations, assumed to be continuous.}

\item{szero}{numeric vector; surrogate marker for control observations, assumed to be continuous.}

\item{yone}{numeric vector; primary outcome for treated observations, assumed to be continuous.}

\item{yzero}{numeric vector; primary outcome for control observations, assumed to be continuous.}

\item{wone}{(only for IPW) numeric vector; probabilities of having non-missing surrogates for treated observations, if \code{NULL} (the default) then SMLE will be used.}

\item{wzero}{(only for IPW) numeric vector; probabilities of having non-missing surrogates for control observations, if \code{NULL} (the default) then SMLE will be used.}

\item{type}{string specifying the type of estimation; choices are \code{"robust"} (the default) or \code{"model"}.}

\item{max.it}{(only for SMLE) scalar; maximum number of iterations allowed in the EM algorithm, the default is \code{1E4}.}

\item{tol}{(only for SMLE) scalar; tolerance between iterations in the EM algorithm used to define convergence, the default is \code{1E-3}.}

\item{conf.int}{logical; if \code{TRUE}, variance estimates and 95% confidence intervals are included in output; if \code{FALSE} (the default), only point estimates are.}

\item{ipw.formula}{(only for IPW) formula; if \code{conf.int = TRUE} and using IPW, the logistic regression formula needed to recalculate the probabilities of non-missing surrogates are needed.}

\item{orig.smle}{(only for SMLE) logical; if \code{TRUE}, SMLE estimated using original code (the default); if \code{FALSE}, it is estimated using code debugged with ChatGPT.}
}
\value{
A list with the following elements:
\describe{
  \item{delta}{Estimated average treatment effect on the primary outcome.}
  \item{delta.s}{Estimated average treatment effect on the surrogate marker.}
  \item{R.s}{Estimated proportion of treatment effect on the primary outcome explained by the treatment effect on a surrogate marker.}
  \item{delta.var}{Estimated variance for \code{delta}.}
  \item{delta.s.var}{Estimated variance for \code{delta.s}.}
  \item{R.s.var}{Estimated variance for \code{R.s}.}
  \item{conf.int.normal.delta}{95\% confidence interval for \code{delta} based on the normal approximation.}
  \item{conf.int.quantile.delta}{95\% confidence interval for \code{delta} based on quantiles.}
  \item{conf.int.normal.delta.s}{95\% confidence interval for \code{delta.s} based on the normal approximation.}
  \item{conf.int.quantile.delta.s}{95\% confidence interval for \code{delta.s} based on quantiles.}
  \item{conf.int.normal.R.s}{95\% confidence interval for \code{R.s} based on the normal approximation.}
  \item{conf.int.quantile.R.s}{95\% confidence interval for \code{R.s} based on quantiles.}
}
}
\description{
Surrogate marker validation with missing values
This function calculates the proportion of treatment effect on the primary outcome explained by the treatment effect on a surrogate marker, correcting for missing values in the surrogate marker. This function is intended to be used for a fully observed continuous outcome. The user must specify what type of estimation they would like (parametric or nonparametric estimation of the proportion explained, denoted by R) and what estimator they would like (see below for details). Missing values can be handled through inverse probability weighting (IPW) or semiparametric sieve maximum likelihood estimation (SMLE).
}
